chore: move security check a few lines up
(for basic.vote contract)

The bug: Currently a null reference Exception can happen because we load

a Account record based on the delegate we want to vote, but when this

Delegate does not exists, then also no Account record exists for this

Delegate. Therefore it would result in a null reference exception.

That's the reason why we move the check for a Delegate a few lines up